Aggregate weights and quantities of materials with same node ID

+++ Models/DTO/ForsMaterial.cs
@@ -7,7 +7,7 @@
 
         public required string MaterialClass { get; init; }
 
-        public double Quantity { get; init; }
+        public int Quantity { get; init; }
 
         public required string MU { get; init; }
